+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Add OSMESA status as "Last Refreshed" to pages using OSMESA stats
 - On user page and dashboard, top campaigns will now show names of campaigns contributed to instead of hashtags contributed to
 - The current version number of Scoreboard is now displayed in the footer
 - On Campaigns Page, "top 10" user table has more detailed statistics